Polymarket_Claude/Claude/requirements.txt
py-clob-client==0.34.6
web3==6.14.0
aiohttp>=3.9.0
aiofiles>=23.0.0
python-telegram-bot==20.7
aiosqlite>=0.20.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tenacity>=8.2.0
pandas>=2.0.0
tabulate>=0.9.0
colorlog>=6.8.0
anthropic>=0.40.0
requests>=2.31.0
